third_party/WebKit/LayoutTests/bluetooth/getPrimaryServices/gen-service-garbage-collection-ran-during-success.html

<!-- Generated by //third_party/WebKit/LayoutTests/bluetooth/generate.py -->
<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../../resources/bluetooth/bluetooth-helpers.js"></script>
<script>
'use strict';
promise_test(() => {
  let promise;
  return setBluetoothFakeAdapter('DisconnectingHealthThermometerAdapter')
    .then(() => requestDeviceWithKeyDown({
      filters: [{services: ['health_thermometer']}]}))
    .then(device => device.gatt.connect())
    .then(gattServer => {
      promise = assert_promise_rejects_with_message(
        gattServer.getPrimaryServices(),
        new DOMException(
          'GATT Server disconnected while retrieving services.',
          'NetworkError'));
      gattServer.disconnect();
    })
    .then(runGarbageCollection)
    .then(() => promise);
}, 'Garbage Collection ran during a getPrimaryServices call that succeeds. ' +
   'Should not crash.');

</script>
